Blocks CRUD Operations v0.8
<details>
 - added Diagramm Relation Block Schema
 - added Diagramm Relation Block Resolver
 - change Diagramm Relation Block Model

+++ backend/src/models/blocks/diagramm_relation_block.js
@@ -6,8 +6,8 @@ export default (sequelize, DataTypes) => {
             primaryKey: true,
         },
         name: DataTypes.STRING,
-        xBlock: DataTypes.DOUBLE,
-        yBlock: DataTypes.DOUBLE,
+        xBlock: DataTypes.FLOAT,
+        yBlock: DataTypes.FLOAT,
         idDiagramm: {
             type: DataTypes.BIGINT,
             allowNull: false,
